Dodano założenie do metody newtona sprawdzajace znak pochodnych

+++ main.py
@@ -1,9 +1,6 @@
-from functions import funtions_value
-from functions import funtions_pattern
 from methods import bisection
 from methods import newton
 from plots import show_plot
-import sympy as sp
 
 # =========================== #
 #   Grzegorz Kucharski 229932 #
@@ -13,7 +10,7 @@ import sympy as sp
 
 while True:
     correct = False
-    function_number = 0
+    function_number = 0  # numer wybranej funkcji
     while correct is not True:
         print("""\nWybierz numer funkcji:
         1 - wielomianowa: 5x^3 + 2x^2 - 1x - 1
@@ -36,8 +33,8 @@ while True:
     while correct is not True:
         print("\nOkresl przedział poszukiwań")
         try:
-            left = int(input("Podaj lewy przedzial: "))
-            right = int(input("Podaj prawy przedzial: "))
+            left = float(input("Podaj lewy przedzial: "))
+            right = float(input("Podaj prawy przedzial: "))
             if left > right:
                 print("Bledny przedzial")
             else:
@@ -69,8 +66,8 @@ while True:
         iteration = int(input("Podaj liczbe iteracji: "))
         accuracy = -1
 
-    result_bisection = bisection(left, right, accuracy, iteration, function_number)
-    result_newton = newton(left, right, accuracy, iteration, function_number)
+    result_bisection = bisection(left, right, accuracy, iteration, function_number)  # miejsce zerowe (bisekcja)
+    result_newton = newton(left, right, accuracy, iteration, function_number)        # miejsce zerowe (netwon)
 
     if result_bisection is False:
         print("Bisekcja: Funkcja nie spełnia założeń na danym przedziale")
@@ -82,4 +79,4 @@ while True:
     else:
         print("Newton - " + str(result_newton))
 
-    show_plot(left, right, function_number, result_bisection, result_newton)
+    show_plot(left, right, function_number, result_bisection, result_newton)  # rysowanie wykresu
